Added eeprom driver

+++ src/keil/drivers/i2c.h
@@ -11,9 +11,9 @@ typedef enum I2CDevice {
 
 void i2cInit(void);
 bool i2cWrite(I2CDevice device, uint8_t addr_, uint8_t reg_, uint8_t data);
-bool i2cWriteBuffer(I2CDevice device, uint8_t addr_, uint8_t reg_, uint8_t len_, uint8_t *data);
-bool i2cRead(I2CDevice device, uint8_t addr_, uint8_t reg_, uint8_t len_, uint8_t *buf);
-bool i2cReadBuffer(I2CDevice device, uint8_t addr_, uint8_t reg_, uint8_t len_, uint8_t *buf);
+bool i2cWriteBuffer(I2CDevice device, uint8_t addr_, uint8_t reg_, uint16_t len_, uint8_t *data);
+bool i2cRead(I2CDevice device, uint8_t addr_, uint8_t reg_, uint16_t len_, uint8_t *buf);
+bool i2cReadBuffer(I2CDevice device, uint8_t addr_, uint8_t reg_, uint16_t len_, uint8_t *buf);
 bool i2cBusy(I2CDevice device, bool *error);
 
 uint16_t i2cGetErrorCounter(void);
